Fix date handling.
